requirements.txt
numpy==1.17.4
opencv-python-headless==4.5.5.62
Pillow==8.4.0; python_version <= '3.6.9'
Pillow==9.0.0; python_version > '3.6.9'
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.2+cu113
torchvision==0.11.3+cu113
-f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html
detectron2==0.6+cu113
PyYAML==5.3.1
